requirements.txt
openai==0.27.4
tiktoken==0.4.0
gradio==3.32.0
langchain==0.0.194
python-dotenv==1.0.0
googletrans==4.0.0rc1
beautifulsoup4==4.11.2
requests==2.28.2
langdetect==1.0.9
chromadb==0.3.22
